+++ lprprinterhandler.h
@@ -1,0 +1,35 @@
+#ifndef LPRPRINTERHANDLER_H
+#define LPRPRINTERHANDLER_H
+
+#include "proofhardware/lprprinter/lprprinter.h"
+
+#include <QObject>
+#include <QString>
+#include <QFile>
+
+#include <functional>
+
+class LprPrinterHandler : public QObject
+{
+    Q_OBJECT
+public:
+    explicit LprPrinterHandler(const QString &printerName, const QString &printerHost, bool acceptsRaw,
+                               bool acceptsFiles, QObject *parent = nullptr);
+
+    using ResultCallback = std::function<void (bool, const QString &)>;
+
+    void printerStatus(ResultCallback callback);
+    void printRaw(const QByteArray &label, ResultCallback callback);
+    void printFile(QSharedPointer<QFile> file, unsigned int quantity, ResultCallback callback);
+
+    bool acceptsRaw() const;
+    bool acceptsFiles() const;
+
+private:
+    Proof::Hardware::LprPrinter *m_printer;
+    QString m_lastError;
+    bool m_acceptsRaw;
+    bool m_acceptsFiles;
+};
+
+#endif // LPRPRINTERHANDLER_H